// green-campus-web-mobile-main/green-campus-web-mobile-main/web/src/app/data/models/login-response.model.ts
export interface LoginResponse {
  jwt: string;
  email: string;
  role: string;
  profileImageUrl: string;
  firstName: string;
  lastName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface ForgotPasswordRequest {
  email: string;
}

export interface ResetPasswordRequest {
  email: string;
  otp: string;
  newPassword: string;
}